Add Shutdown state and update map rendering logic
The `AppStateEnum` enumeration was updated to include a new state called `Shutdown`. The map rendering logic in `Map.razor` was modified to conditionally render the map and other UI elements only if the `AppState` is not `Shutdown`. A new check was added in `Map.razor` to display an error dialog if the `AppState` is `Shutdown`. The `MapStateService` constructor was updated to handle the new `Shutdown` state based on the configuration value. The `appsettings.json` file was updated to include the new `Shutdown` state and added messages for both Czech and English languages to inform users about the application's downtime. A minor change was made in `index.html` to add a space before a script tag.

The `mapsettings.pacov.json` file has been updated to change the initial zoom level from 17 to 15, and the latitude and longitude coordinates have been adjusted slightly. Additionally, the minimum zoom level for "interest" and "address" place types has been changed from 16 to 15. In the `site.ts` file, a minor syntax correction was made to the `heatmapData` type definition by adding a semicolon. Also, a new configuration object for the heatmap was added in the `showHeatmap` function. The `ErrorDialog.razor` file has been significantly updated to include dependency injections for `IStringLocalizer<CommonResources>`, `DialogService`, `IJSRuntime`, `MapStateService`, and `IConfiguration`. The UI has been enhanced to display different messages and styles based on whether the view is mobile or not. For non-mobile views, it shows a message indicating that "MemoMap is down" in both Czech and English. For mobile views, it shows a simplified message with the app state.